+++ services/libs/data-access-layer/src/osspckgs/index.ts
@@ -3,5 +3,6 @@ export * from './packages'
 export * from './maintainers'
 export * from './versions'
 export * from './repos'
+export * from './sqlFragments'
 export * from './stewardships'
 export * from './api'
